fix: big fixes (15 tests left)

+++ src/main/java/ru/practicum/shareit/request/dao/RequestDao.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.practicum.shareit.request.model.ItemRequest;
 
 @Repository
-public interface RequestDao extends JpaRepository<ItemRequest, Integer> {
+public interface RequestDao extends JpaRepository<ItemRequest, Long> {
 }
